fix: split modules & functions
